requirements.txt
torch-ema
ninja
trimesh
opencv-python
tensorboardX
numpy==1.23.5
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn
flask
face_alignment
python_speech_features
numba
resampy
#pyaudio
soundfile
einops
configargparse
google-generativeai
openai
modelscope
lpips
imageio-ffmpeg

transformers==4.30.0
edge_tts==6.1.11
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate

librosa
